domain/src/main/java/com/yuryhalubets/domain/use_case/get_coins/GetCoinsUseCaseImpl.kt
/*
 * Copyright (C) 2022, Yury Halubets
 * All rights reserved.
 */
package com.yuryhalubets.domain.use_case.get_coins

import com.yuryhalubets.core.domain.Resource
import com.yuryhalubets.domain.models.Coin
import com.yuryhalubets.domain.repository.CoinRepository
import kotlinx.coroutines.flow.Flow
import kotlinx.coroutines.flow.flow
import java.lang.Exception
import javax.inject.Inject

internal class GetCoinsUseCaseImpl @Inject constructor(
    private val repository: CoinRepository
) : GetCoinsUseCase {
    override operator fun invoke(): Flow<Resource<List<Coin>>> = flow {
        try {
            emit(Resource.Loading())
            val coins = repository.getCoins()
            emit(Resource.Success(coins))
        } catch (e: Exception) {
            emit(Resource.Error(e.localizedMessage ?: "An unexpected error occured"))
        }
    }
}